+++ publish-common/src/main/java/com/ericsson/eiffel/remrem/publish/service/GenerateURLTemplate.java
@@ -1,10 +1,10 @@
-package com.ericsson.eiffel.remrem.publish.controller;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
+package com.ericsson.eiffel.remrem.publish.service;
 
 import java.util.HashMap;
 import java.util.Map;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
 
 /**
  * A class for generating URL template of REMReM Generate Service
@@ -19,7 +19,23 @@ public class GenerateURLTemplate {
     @Value("${generate.server.path}")
     private String generateServerPath;
 
-    public String getUrl() {
+	public String getGenerateServerUri() {
+		return generateServerUri;
+	}
+
+	public void setGenerateServerUri(String generateServerUri) {
+		this.generateServerUri = generateServerUri;
+	}
+
+	public String getGenerateServerPath() {
+		return generateServerPath;
+	}
+
+	public void setGenerateServerPath(String generateServerPath) {
+		this.generateServerPath = generateServerPath;
+	}
+
+	public String getUrl() {
         return generateServerUri + generateServerPath + "/{mp}?msgType={msgType}";
     }