+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.5
+- Fixed generateAndPublish endpoint.
+- Added functionality to load Generate server properties from JAVA_OPTS.
+
 ## 1.0.4
 - Fixed publish-CLI by making necessary code change in publish-common.